Move manifest into a package. Provide an interface for it that is currently not used.
