+++ pkg/manifest/manifest_test.go
@@ -1,4 +1,4 @@
-package canary
+package manifest
 
 import (
 	"fmt"
